requirements.txt
requests
beautifulsoup4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
importlib